refactor: consolidate test files
