--- packages/e2e-tests/plugins/interactive-blocks/render-html-array/block.json
@@ -1,15 +1,0 @@
-{
-	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 3,
-	"name": "test/render-html-array",
-	"title": "E2E Interactivity tests - render html array",
-	"category": "text",
-	"icon": "heart",
-	"description": "",
-	"supports": {
-		"interactivity": true
-	},
-	"textdomain": "e2e-interactivity",
-	"viewScriptModule": "file:./view.js",
-	"render": "file:./render.php"
-}
