--- packages/e2e-tests/plugins/interactive-blocks/render-html-array/view.asset.php
@@ -1,5 +1,0 @@
-<?php return array(
-	'dependencies' => array(
-		'@wordpress/interactivity',
-	),
-);
