--- packages/e2e-tests/plugins/interactive-blocks/render-html-region/view.asset.php
@@ -1,1 +1,0 @@
-<?php return array( 'dependencies' => array( '@wordpress/interactivity' ) );
