--- packages/e2e-tests/plugins/interactive-blocks/render-html-region/view.js
@@ -1,8 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { store } from '@wordpress/interactivity';
-
-// Register the same store used by the `test/render-html` block, so the
-// region's directives resolve against it.
-store( 'test/render-html', {} );
